Add a getauxval backend
