+++ include/xsimd/config/xsimd_cpu_features_arm.hpp
@@ -13,11 +13,9 @@
 #define XSIMD_CPU_FEATURES_ARM_HPP
 
 #include "./xsimd_config.hpp"
-
-#if XSIMD_TARGET_ARM && XSIMD_HAVE_LINUX_GETAUXVAL
-#include "../utils/bits.hpp"
 #include "./xsimd_getauxval.hpp"
 
+#if XSIMD_TARGET_ARM && XSIMD_HAVE_LINUX_GETAUXVAL
 // HWCAP_XXX masks to use on getauxval results.
 // Header does not exists on all architectures and masks are architecture
 // specific.
@@ -36,7 +34,7 @@ namespace xsimd
      * This is well defined on all architectures.
      * It will always return false on non-ARM architectures.
      */
-    class arm_cpu_features
+    class arm_cpu_features : private linux_hwcap_backend_default
     {
     public:
         arm_cpu_features() noexcept = default;
@@ -78,45 +76,6 @@ namespace xsimd
             return false;
 #endif
         }
-
-    private:
-#if XSIMD_TARGET_ARM && XSIMD_HAVE_LINUX_GETAUXVAL
-        enum class status
-        {
-            hwcap_valid = 0,
-            hwcap2_valid = 1,
-        };
-
-        using status_bitset = utils::uint_bitset<status, std::uint32_t>;
-
-        mutable status_bitset m_status {};
-
-        mutable xsimd::linux_auxval m_hwcap {};
-
-        inline xsimd::linux_auxval const& hwcap() const noexcept
-        {
-            if (!m_status.bit_is_set<status::hwcap_valid>())
-            {
-                m_hwcap = xsimd::linux_auxval::read(AT_HWCAP);
-                m_status.set_bit<status::hwcap_valid>();
-            }
-            return m_hwcap;
-        }
-
-#if XSIMD_TARGET_ARM64
-        mutable xsimd::linux_auxval m_hwcap2 {};
-
-        inline xsimd::linux_auxval const& hwcap2() const noexcept
-        {
-            if (!m_status.bit_is_set<status::hwcap2_valid>())
-            {
-                m_hwcap2 = xsimd::linux_auxval::read(AT_HWCAP2);
-                m_status.set_bit<status::hwcap2_valid>();
-            }
-            return m_hwcap2;
-        }
-#endif
-#endif // XSIMD_TARGET_ARM && XSIMD_HAVE_LINUX_GETAUXVAL
     };
 }
 #endif
