+++ include/xsimd/config/xsimd_cpu_features_riscv.hpp
@@ -13,11 +13,9 @@
 #define XSIMD_CPU_FEATURES_RISCV_HPP
 
 #include "./xsimd_config.hpp"
-
-#if XSIMD_TARGET_RISCV && XSIMD_HAVE_LINUX_GETAUXVAL
-#include "../utils/bits.hpp"
 #include "./xsimd_getauxval.hpp"
 
+#if XSIMD_TARGET_RISCV && XSIMD_HAVE_LINUX_GETAUXVAL
 // HWCAP_XXX masks to use on getauxval results.
 // Header does not exists on all architectures and masks are architecture
 // specific.
@@ -26,7 +24,7 @@
 
 namespace xsimd
 {
-    class riscv_cpu_features
+    class riscv_cpu_features : private linux_hwcap_backend_default
     {
     public:
         riscv_cpu_features() noexcept = default;
@@ -44,30 +42,6 @@ namespace xsimd
             return false;
 #endif
         }
-
-    private:
-#if XSIMD_TARGET_RISCV && XSIMD_HAVE_LINUX_GETAUXVAL
-        enum class status
-        {
-            hwcap_valid = 0,
-        };
-
-        using status_bitset = utils::uint_bitset<status, std::uint32_t>;
-
-        mutable status_bitset m_status {};
-
-        mutable xsimd::linux_auxval m_hwcap {};
-
-        inline xsimd::linux_auxval const& hwcap() const noexcept
-        {
-            if (!m_status.bit_is_set<status::hwcap_valid>())
-            {
-                m_hwcap = xsimd::linux_auxval::read(AT_HWCAP);
-                m_status.set_bit<status::hwcap_valid>();
-            }
-            return m_hwcap;
-        }
-#endif
     };
 }
